+++ src/design/plone/policy/restapi/bandi_search_filters/get.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from AccessControl.unauthorized import Unauthorized
 from plone import api
 from plone.restapi.services import Service
 from zope.interface import implementer
@@ -9,32 +8,20 @@ from zope.publisher.interfaces import IPublishTraverse
 @implementer(IPublishTraverse)
 class BandiSearchFiltersGet(Service):
     def reply(self):
-        subjects_keys = []
-        offices_keys = []
-        subjects = []
-        offices = []
+        subjects = [
+            {"UID": x, "title": x}
+            for x in self.context.portal_catalog.uniqueValuesFor("Subject_bando")
+        ]
 
-        # populate with only list from visible bandi
-        for brain in api.content.find(portal_type="Bando"):
-            for subject in getattr(brain, "Subject_bando", []) or []:
-                if subject in subjects_keys:
-                    continue
-                subjects_keys.append(subject)
-                subjects.append({"UID": subject, "title": subject})
-            for uid in getattr(brain, "ufficio_responsabile_bando", []) or []:
-                if uid in offices_keys:
-                    continue
-                # add also if it's private, so we will not check it anymore
-                offices_keys.append(uid)
-                try:
-                    item = api.content.get(UID=uid)
-                except Unauthorized:
-                    continue
-                if item:
-                    offices.append({"UID": uid, "title": item.Title()})
-        offices.sort(key=lambda x: x["title"])
+        uffici_uids = self.context.portal_catalog.uniqueValuesFor(
+            "ufficio_responsabile_bando"
+        )
+        offices = [
+            {"UID": x.UID, "title": x.Title} for x in api.content.find(UID=uffici_uids)
+        ]
+
         subjects.sort(key=lambda x: x["title"])
-
+        offices.sort(key=lambda x: x["title"])
         return {
             "subjects": subjects,
             "offices": offices,